pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrheem"
version = "0.1.0"
description = "Python library and CLI for controlling Rheem water heaters via the EcoNet API"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Riley Porter", email = "rileyporter@gmail.com"}
]
keywords = ["rheem", "econet", "water heater", "iot", "smart home", "home automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "paho-mqtt>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "python-dotenv>=0.19.0",
]

[project.scripts]
rheem = "pyrheem.cli:main"

[project.urls]
Homepage = "https://github.com/rileyporter/pyrheem"
Issues = "https://github.com/rileyporter/pyrheem/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pyrheem*"]

[tool.ruff]
target-version = "py38"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "F401", "SIM102", "SIM108"]

[tool.ruff.format]
quote-style = "double"
